api/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-jose==3.3.0
PyJWT==2.8.0
bcrypt==4.0.1 